add layout component
