+++ 1 - Students/ArtemKarlov/package.json
@@ -27,6 +27,7 @@
     "@material-ui/icons": "^4.9.1",
     "prop-types": "^15.7.2",
     "react": "^16.13.1",
-    "react-dom": "^16.13.1"
+    "react-dom": "^16.13.1",
+    "react-router-dom": "^5.2.0"
   }
 }
