+++ 1 - Students/ArtemKarlov/src/components/Layout/style.css
@@ -1,0 +1,52 @@
+* {
+    margin: 0;
+    padding: 0;
+}
+
+body {
+    font-family: Arial, Helvetica, sans-serif;
+    font-size: 12px;
+    font-weight: normal;
+    font-style: normal;
+    line-height: 1.2;
+    color: #FFFFFF;
+    background-color: #e0e0e0;
+}
+
+.layout {
+    max-width: 768px;
+    height: 90vh;
+    min-height: 450px;
+    margin-top: 40px;
+    margin: 0 auto;
+    display: flex;
+    flex-direction: row;    
+    /* background-color: brown; */
+}
+
+.layout__control-panel {
+    flex-shrink: 0;
+    flex-grow: 0;
+}
+
+.layout__account-chatlist {
+    width: 270px;
+    height: 100%;
+
+    flex-shrink: 0;
+    flex-grow: 0;
+
+
+    display: flex;
+    flex-direction: column;
+    flex-wrap: nowrap;
+}
+
+.layout__chats {
+    flex-grow: 1;
+}
+
+.lyaout__chat {
+    flex-grow: 1;
+    flex-shrink: 1;
+}
